requirements.txt
streamlit>=1.24.0
langchain>=0.0.350
openai>=1.3.0
psycopg2-binary>=2.9.9
beautifulsoup4>=4.12.2
google-search-results>=2.4.2
requests>=2.31.0
python-dotenv>=1.0.0
pypdf>=3.17.1
weasyprint>=60.1
markdown>=3.4.3
numpy>=1.24.0
pgvector>=0.2.3
langchain-community>=0.0.28
langchain-openai>=0.2.14
tiktoken>=0.5.1
python-markdown-math>=0.8
azure-cognitiveservices-speech>=1.31.0
reportlab>=4.0.0
